Added better error display when git directory is not accessible

+++ src/main/java/com/greendelta/collaboration/service/RepositoryService.java
@@ -20,7 +20,6 @@ import org.apache.logging.log4j.Logger;
 import org.eclipse.jgit.api.Git;
 import org.eclipse.jgit.api.ResetCommand;
 import org.eclipse.jgit.api.ResetCommand.ResetType;
-import org.eclipse.jgit.api.errors.GitAPIException;
 import org.eclipse.jgit.internal.storage.file.FileRepository;
 import org.openlca.git.model.Commit;
 import org.openlca.jsonld.LibraryLink;
@@ -31,6 +30,7 @@ import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 import org.springframework.web.servlet.mvc.method.annotation.StreamingResponseBody;
 
+import com.greendelta.collaboration.controller.util.Response;
 import com.greendelta.collaboration.error.ForbiddenAccessException;
 import com.greendelta.collaboration.error.RepositoryNotFoundException;
 import com.greendelta.collaboration.error.UnsupportedSchemaException;
@@ -135,19 +135,21 @@ public class RepositoryService {
 			throw new ForbiddenAccessException(group, "WRITE");
 		var path = getPath(group, name);
 		if (path == null)
-			throw new ForbiddenAccessException(group, "WRITE");
-		init(path);
+			return null;
+		if (!init(path))
+			return null;
 		membershipService.addMembership(currentUser, RepositoryPath.of(group, name).toString(), Role.OWNER, true);
 		return get(group, name);
 	}
 
-	private void init(String path) {
+	private boolean init(String path) {
 		File dir = new File(path);
 		try (var git = Git.init().setBare(true).setDirectory(dir).call()) {
-		} catch (GitAPIException e) {
+			return true;
+		} catch (Exception e) {
 			log.error("Error initializing git repository", e);
 			Dirs.delete(dir);
-			throw new Error(e);
+			return false;
 		}
 	}
 
@@ -161,6 +163,8 @@ public class RepositoryService {
 		if (exists(group, name))
 			return false;
 		try (var newRepo = create(group, name)) {
+			if (newRepo == null)
+				throw Response.error("Could not create repository, does the configured 'Repositories root directory' exist and can be write-accessed?");
 			Dirs.move(repo.dir.toPath(), newRepo.dir.toPath());
 			moveMemberships(repo, newRepo);
 			commentService.move(repo, newRepo);
@@ -254,6 +258,8 @@ public class RepositoryService {
 
 	public void unpack(Repository repo, InputStream input) {
 		try {
+			if (!repo.dir.exists())
+				return;
 			Dirs.delete(repo.dir);
 			create(repo.group, repo.name).close();
 			var repoPath = repo.dir.toPath();
